215: move Configuration File away from access -> also has UI specific values move ICON_FONT_SIZE to config
