+++ src/main/java/gvs/business/logic/layouter/tree/TreeLayouter.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 
 import com.google.common.collect.Lists;
 
-import gvs.access.Configuration;
+import gvs.Configuration;
 import gvs.business.logic.Session;
 import gvs.business.logic.layouter.ILayouter;
 import gvs.business.model.Graph;
