+++ src/main/java/gvs/business/model/tree/TreeVertex.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Observable;
 
-import gvs.access.Configuration;
+import gvs.Configuration;
 import gvs.business.model.IVertex;
 import gvs.business.model.styles.GVSStyle;
 import gvs.util.FontAwesome.Glyph;
