+++ src/main/java/gvs/ui/model/VertexViewModel.java
@@ -6,6 +6,7 @@ import java.util.Observer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import gvs.Configuration;
 import gvs.ScalablePane;
 import gvs.business.model.IVertex;
 import gvs.business.model.styles.GVSStyle;
@@ -34,8 +35,6 @@ public class VertexViewModel implements Observer {
   private final Label label;
   private final Ellipse ellipse;
 
-  private static final int ICON_FONT_SIZE = 8;
-
   private static final Logger logger = LoggerFactory
       .getLogger(VertexViewModel.class);
 
@@ -60,8 +59,8 @@ public class VertexViewModel implements Observer {
     // setup label text and optional icon
     if (vertex.getIcon() != null) {
       logger.info("Creating VertexViewModel with an icon");
-      label.setGraphic(
-          FontAwesome.createLabel(vertex.getIcon(), ICON_FONT_SIZE));
+      label.setGraphic(FontAwesome.createLabel(vertex.getIcon(),
+          Configuration.getIconFontSize()));
     }
     label.setText(vertex.getLabel());
     if (!vertex.isTreeVertex()) {
